+++ num/NumType/test/NumType_test.cpp
@@ -7,11 +7,11 @@ using alg::num::NumType;
 
 TEST(NumType, RealReal_OperatorA_test)
 {
-    Real<long double> x1(256);
-    Real<long double> x2(32);
+    Real<double> x1(256);
+    Real<double> x2(32);
 
-    NumType<long double> var1 = x1;
-    NumType<long double> var2 = x2;
+    NumType<double> var1 = x1;
+    NumType<double> var2 = x2;
 
     EXPECT_DOUBLE_EQ((var1+var2).getReal().real(), (x1 + x2).real());
     EXPECT_DOUBLE_EQ((var1-var2).getReal().real(), (x1 - x2).real());
@@ -21,11 +21,11 @@ TEST(NumType, RealReal_OperatorA_test)
 
 TEST(NumType, RealComplex_OperatorA_test)
 {
-    Real<long double> x1_real(256); Complex<long double> x1(256);
-    Complex<long double> x2(32, 2);
+    Real<double> x1_real(256); Complex<double> x1(256);
+    Complex<double> x2(32, 2);
 
-    NumType<long double> var1 = x1_real;
-    NumType<long double> var2 = x2;
+    NumType<double> var1 = x1_real;
+    NumType<double> var2 = x2;
 
     EXPECT_DOUBLE_EQ((var1+var2).getComplex().real(), (x1 + x2).real());
     EXPECT_DOUBLE_EQ((var1+var2).getComplex().imag(), (x1 + x2).imag());
@@ -39,11 +39,11 @@ TEST(NumType, RealComplex_OperatorA_test)
 
 TEST(NumType, ComplexComplex_OperatorA_test)
 {
-    Complex<long double> x1(256, 50);
-    Complex<long double> x2(32, 25);
+    Complex<double> x1(256, 50);
+    Complex<double> x2(32, 25);
 
-    NumType<long double> var1 = x1;
-    NumType<long double> var2 = x2;
+    NumType<double> var1 = x1;
+    NumType<double> var2 = x2;
 
     EXPECT_DOUBLE_EQ((var1+var2).getComplex().real(), (x1 + x2).real());
     EXPECT_DOUBLE_EQ((var1+var2).getComplex().imag(), (x1 + x2).imag());
